pyproject.toml
[project]
name = "semantic-video-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "faster-whisper>=1.2.1",
    "ffmpeg-python>=0.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-google-genai>=4.2.1",
    "numpy>=2.2.6",
    "pydantic>=2.12.5",
    "pyttsx3>=2.99",
    "sentence-transformers>=5.2.3",
    "speechrecognition>=3.14.5",
    "yt-dlp>=2026.2.21",
]
